graphrag_mcp_py/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
pydantic>=2.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
python-multipart>=0.0.5
requests>=2.28.0
pytest>=7.0.0